backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
google-generativeai==0.8.3
gspread==6.1.4
python-dotenv==1.0.1
pydantic==2.10.6
pydantic-settings==2.7.1
google-auth==2.37.0
google-auth-oauthlib==1.2.1
bleach==6.1.0
email-validator==2.2.0
email-validator==2.2.0